+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Bug fixes:
 
 Compatibility:
 
+* Fix `MatchData#[]` when passed unbounded Range (#2755, @andrykonchin).
 
 Performance: